#23 ref(image): Split into component the image card
